Move icon/gravity_waves/ to gravity_waves/icon/
Convert relative imports to absolute, update icon_physics.py and
parameters.py. Make icon/__init__.py use lazy imports to avoid
circular dependencies from the reorganization.
