+++ jcm/physics/gravity_waves/icon/gravity_wave_drag_test.py
@@ -9,7 +9,7 @@ from .gravity_wave_drag import (
     GravityWaveParameters, brunt_vaisala_frequency, orographic_source, wave_breaking_criterion,
     gravity_wave_drag
 )
-from ..constants.physical_constants import grav, cp
+from jcm.physics.icon.constants.physical_constants import grav, cp
 
 
 class TestBruntVaisalaFrequency:
