+++ jcm/physics/icon/parameters.py
@@ -10,7 +10,7 @@ import tree_math
 
 from .convection import ConvectionParameters
 from .clouds import CloudParameters, MicrophysicsParameters
-from .gravity_waves import GravityWaveParameters
+from jcm.physics.gravity_waves.icon import GravityWaveParameters
 from .radiation import RadiationParameters
 from .vertical_diffusion.vertical_diffusion_types import VDiffParameters
 from .surface import SurfaceParameters
